Part 2: Create Unit Tests
Part 2: Create Unit Tests
https://www.youtube.com/watch?v=whhssgDHRP0
Create some sample unit tests for the ProcessTrades class. It is OK if some of this test fail for now. We will create more tests in the assignment.
First, try to create unit tests for ReadTradeData() in the TradeProcessor class. This should not work and you should see the error that you can only create unit tests for public methods and ReadTradeData() is private.
Make a unit tests for the public ProcessTrades() method. You will probably want to use the CountDbRecords() method--remember to change the connection string here and add a reference to System.Data
See this Gist sample code for an example of a Unit Test.
You should also add some sample text files for the trades to the text project. The easiest way to do this is to copy the trades.txt file from the main project into the test project
For each text file you add, check that its Build property to “Embedded Resource”